do a bit of refactoring
