+++ src/Maestro/Maestro.Web/Startup.cs
@@ -248,7 +248,7 @@ public partial class Startup : StartupBase
         services.AddTransient<IProcessManager>(sp =>
             new ProcessManager(
                 sp.GetRequiredService<ILogger<ProcessManager>>(),
-                "git"));
+                LocalGit.GetPathToLocalGit()));
         services.AddTransient<IVersionDetailsParser, VersionDetailsParser>();
         services.AddScoped<IRemoteFactory, DarcRemoteFactory>();
         services.AddTransient<IBasicBarClient, SqlBarClient>();
