+++ src/Microsoft.DotNet.Darc/DarcLib/Helpers/LocalGit.cs
@@ -1,19 +1,14 @@
-// Licensed to the .NET Foundation under one or more agreements.
+﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System;
 using System.IO;
+using System;
 
-namespace SubscriptionActorService;
+namespace Microsoft.DotNet.DarcLib.Helpers;
 
-public interface ILocalGit
+public static class LocalGit
 {
-    string GetPathToLocalGit();
-}
-
-public class LocalGit : ILocalGit
-{
-    public string GetPathToLocalGit()
+    public static string GetPathToLocalGit()
     {
         var gitExePath = Path.Join(AppContext.BaseDirectory, "git-portable", "bin", "git.exe");
         if (!File.Exists(gitExePath))
